+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/Catalog/SceneObjectDropController.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.EventSystems;
+
+public class SceneObjectDropController : MonoBehaviour
+{
+    public CatalogGroupListView catalogGroupListView;
+    public event Action<SceneObject> OnSceneObjectDropped;
+
+    public void SceneObjectDropped(BaseEventData data)
+    {
+        CatalogItemAdapter adapter = catalogGroupListView.GetLastSceneObjectDragged();
+        SceneObject sceneObject = adapter.GetContent();
+
+        OnSceneObjectDropped?.Invoke(sceneObject);
+    }
+}
