+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/BuildModeHUD/Resources/BuildModeHUD.prefab
@@ -1799,7 +1799,7 @@ RectTransform:
   m_Children:
   - {fileID: 851071763565322553}
   m_Father: {fileID: 851071764747098679}
-  m_RootOrder: 5
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 1, y: 1}
@@ -4848,6 +4848,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   contentPanelTransform: {fileID: 851071764701594716}
+  generalCanvas: {fileID: 851071764747098676}
   categoryItemAdapterPrefab: {fileID: 751415691112004589, guid: 6bcacb01925475b49b2fd03bfe1fd433,
     type: 3}
 --- !u!114 &851071763915059373
@@ -5251,7 +5252,7 @@ RectTransform:
   m_Children:
   - {fileID: 851071764148484467}
   m_Father: {fileID: 851071764747098679}
-  m_RootOrder: 6
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -5476,7 +5477,7 @@ RectTransform:
   m_Children:
   - {fileID: 851071763556904382}
   m_Father: {fileID: 851071764747098679}
-  m_RootOrder: 10
+  m_RootOrder: 11
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 1, y: 1}
@@ -11770,12 +11771,12 @@ RectTransform:
   m_Children:
   - {fileID: 851071763810502910}
   m_Father: {fileID: 851071764747098679}
-  m_RootOrder: 8
+  m_RootOrder: 9
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0.5, y: 0}
-  m_AnchorMax: {x: 0.5, y: 0}
-  m_AnchoredPosition: {x: 0, y: 450}
-  m_SizeDelta: {x: 1440, y: 900}
+  m_AnchorMin: {x: 0, y: 0}
+  m_AnchorMax: {x: 1, y: 1}
+  m_AnchoredPosition: {x: 0, y: 0}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &2858140822079303849
 MonoBehaviour:
@@ -14088,6 +14089,7 @@ RectTransform:
   - {fileID: 851071764308614461}
   - {fileID: 851071763631578194}
   - {fileID: 851071764161517832}
+  - {fileID: 84112946647708823}
   - {fileID: 851071765481101389}
   - {fileID: 851071765252996792}
   - {fileID: 851071763679374502}
@@ -14120,6 +14122,7 @@ MonoBehaviour:
   sceneObjectCatalogController: {fileID: 6818990461646912356}
   toolTipController: {fileID: 4730837589599637360}
   quickBarView: {fileID: 2858140822079303849}
+  catalogGroupListView: {fileID: 851071763915059372}
   firstPersonCanvasGO: {fileID: 851071763679374503}
   godModeCanvasGO: {fileID: 851071765252996793}
   extraBtnsGO: {fileID: 851071764922696198}
@@ -19810,7 +19813,7 @@ RectTransform:
   - {fileID: 851071765511369652}
   - {fileID: 851071765682423715}
   m_Father: {fileID: 851071764747098679}
-  m_RootOrder: 4
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 1, y: 1}
@@ -23549,7 +23552,7 @@ RectTransform:
   m_Children:
   - {fileID: 851071765510951995}
   m_Father: {fileID: 851071764747098679}
-  m_RootOrder: 3
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
@@ -24306,7 +24309,7 @@ RectTransform:
   m_Children:
   - {fileID: 851071764805437011}
   m_Father: {fileID: 851071764747098679}
-  m_RootOrder: 7
+  m_RootOrder: 8
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -24792,7 +24795,7 @@ RectTransform:
   m_Children:
   - {fileID: 851071764327199312}
   m_Father: {fileID: 851071764747098679}
-  m_RootOrder: 9
+  m_RootOrder: 10
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 1, y: 1}
@@ -24816,6 +24819,7 @@ MonoBehaviour:
   catalogGroupListView: {fileID: 851071763915059372}
   searchInputField: {fileID: 851071764848031699}
   quickBarView: {fileID: 2858140822079303849}
+  sceneObjectDropController: {fileID: 1480823636982834102}
 --- !u!1 &851071765561143009
 GameObject:
   m_ObjectHideFlags: 0
@@ -26585,6 +26589,123 @@ MonoBehaviour:
   - {fileID: 0}
   m_baseMaterial: {fileID: 0}
   m_maskOffset: {x: 0, y: 0, z: 0, w: 0}
+--- !u!1 &4413728610159217901
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 84112946647708823}
+  - component: {fileID: 1480823636982834102}
+  - component: {fileID: 5887779308624769835}
+  - component: {fileID: 7086200805059309350}
+  - component: {fileID: 5758824516929350152}
+  m_Layer: 5
+  m_Name: DragAndDropSceneObject
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!224 &84112946647708823
+RectTransform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4413728610159217901}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 851071764747098679}
+  m_RootOrder: 3
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+  m_AnchorMin: {x: 0, y: 0}
+  m_AnchorMax: {x: 1, y: 1}
+  m_AnchoredPosition: {x: 0, y: 0}
+  m_SizeDelta: {x: 0, y: 0}
+  m_Pivot: {x: 0.5, y: 0.5}
+--- !u!114 &1480823636982834102
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4413728610159217901}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 224e0c95cd2c7b64d93d9eba824f60e3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  catalogGroupListView: {fileID: 851071763915059372}
+--- !u!222 &5887779308624769835
+CanvasRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4413728610159217901}
+  m_CullTransparentMesh: 0
+--- !u!114 &7086200805059309350
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4413728610159217901}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Material: {fileID: 0}
+  m_Color: {r: 1, g: 1, b: 1, a: 0}
+  m_RaycastTarget: 1
+  m_Maskable: 1
+  m_OnCullStateChanged:
+    m_PersistentCalls:
+      m_Calls: []
+  m_Sprite: {fileID: 0}
+  m_Type: 0
+  m_PreserveAspect: 0
+  m_FillCenter: 1
+  m_FillMethod: 4
+  m_FillAmount: 1
+  m_FillClockwise: 1
+  m_FillOrigin: 0
+  m_UseSpriteMesh: 0
+  m_PixelsPerUnitMultiplier: 1
+--- !u!114 &5758824516929350152
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4413728610159217901}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d0b148fe25e99eb48b9724523833bab1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Delegates:
+  - eventID: 6
+    callback:
+      m_PersistentCalls:
+        m_Calls:
+        - m_Target: {fileID: 1480823636982834102}
+          m_MethodName: SceneObjectDropped
+          m_Mode: 0
+          m_Arguments:
+            m_ObjectArgument: {fileID: 0}
+            m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
+            m_IntArgument: 0
+            m_FloatArgument: 0
+            m_StringArgument: 
+            m_BoolArgument: 0
+          m_CallState: 2
 --- !u!1 &4616652751768278129
 GameObject:
   m_ObjectHideFlags: 0
