--- unity-client/Assets/Scripts/MainScripts/DCL/Controllers/Scene/WorldState.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 97c8f65c1398b3941945c434dac16868
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
